实现相册修改头像并使用Http更新
