+++ app/src/main/java/com/example/bigworks/utils/ImageModify.java
@@ -1,0 +1,177 @@
+package com.example.bigworks.utils;
+
+import android.Manifest;
+import android.annotation.TargetApi;
+import android.app.Activity;
+import android.app.Dialog;
+import android.content.ContentUris;
+import android.content.Context;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.database.Cursor;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.net.Uri;
+import android.os.Build;
+import android.provider.DocumentsContract;
+import android.provider.MediaStore;
+import android.util.Log;
+import android.view.Gravity;
+import android.view.View;
+import android.view.Window;
+import android.view.WindowManager;
+import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.Toast;
+
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.FileProvider;
+
+import com.example.bigworks.R;
+import com.example.bigworks.http.UserData.Http_setHeadImg;
+import com.example.bigworks.persondata.PersonDataPageActivity;
+
+import java.io.File;
+import java.io.IOException;
+
+import static androidx.core.app.ActivityCompat.startActivityForResult;
+import static androidx.core.app.ActivityCompat.startIntentSenderForResult;
+
+public class ImageModify {
+    private Uri imageUri;
+    private Button takePhoto;
+    private Button cancel;
+    private Button chooseFromAlbum;
+    private Context context;
+    private static final int TAKE_PHOTO = 1;
+    private static final int CHOOSE_PHOTO = 2;
+    private static final int CROP_PHOTO = 3;
+
+    //对话框相机
+    public void initImagePhoto(Context context,Dialog dialog){
+        this.context = context;
+        takePhoto = (Button) dialog.findViewById(R.id.image_photo);
+        takePhoto.setOnClickListener(new View.OnClickListener(){
+            public void onClick(View v){
+                File outputImage = new File(context.getExternalCacheDir(), "output_image.jpg");
+                try{
+                    if(outputImage.exists()){
+                        outputImage.delete();
+                    }
+                    outputImage.createNewFile();
+                }catch (IOException e){
+                    e.printStackTrace();
+                }
+                if(Build.VERSION.SDK_INT >= 24){
+                    imageUri = FileProvider.getUriForFile(context, "com.example.bigworks.fileprovider", outputImage);
+                }else{
+                    imageUri = Uri.fromFile(outputImage);
+                }
+                Intent intent = new Intent("android.media.action.IMAGE_CAPTURE");
+                intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
+                startActivityForResult((Activity) context, intent, TAKE_PHOTO, null);
+            }
+        });
+    }
+
+    //对话框取消
+    public void imageCancel(Dialog dialog){
+        cancel = (Button) dialog.findViewById(R.id.image_cancel);
+        cancel.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                dialog.dismiss();
+            }
+        });
+    }
+
+    //对话框相册
+    public void initImageAlbum(Context context, Dialog dialog){
+        this.context = context;
+        chooseFromAlbum = (Button) dialog.findViewById(R.id.image_album);
+        chooseFromAlbum.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED){
+                    ActivityCompat.requestPermissions((Activity) context, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
+                }else{
+                    openAlbum();
+                }
+            }
+        });
+    }
+
+    private void openAlbum(){
+        Intent intent = new Intent("android.intent.action.GET_CONTENT");
+        intent.setType("image/*");
+        startActivityForResult((Activity) this.context, intent, CHOOSE_PHOTO, null);
+    }
+
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults){
+        switch (requestCode){
+            case 1:
+                if(grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
+                    openAlbum();
+                }else{
+                    Toast.makeText(context, "You denide the permission",Toast.LENGTH_SHORT).show();
+                }
+                break;
+            default:
+                break;
+        }
+    }
+
+    @TargetApi(19)
+    public void handleImageOnKitKat(Intent data){
+        String imagePath = null;
+        Uri uri = data.getData();
+        if(DocumentsContract.isDocumentUri(this.context, uri)){
+            String docId = DocumentsContract.getDocumentId(uri);
+            if("com.android.providers.media.documents".equals(uri.getAuthority())){
+                String id = docId.split(":")[1];
+                String selection = MediaStore.Images.Media._ID + "=" + id;
+                imagePath = getImagePath(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, selection);
+            }else if("com.android.providers.downloads.documents".equals(uri.getAuthority())){
+                Uri contentUri = ContentUris.withAppendedId(Uri.parse("content://downloads/public_downloads"), Long.valueOf(docId));
+                imagePath = getImagePath(contentUri, null);
+            }
+        }else if("content".equalsIgnoreCase(uri.getScheme())){
+            imagePath = getImagePath(uri, null);
+        }else if("file".equalsIgnoreCase(uri.getScheme())){
+            imagePath = uri.getPath();
+        }
+        displayImage(imagePath);
+    }
+
+    public void handleImageBeforeKitkat(Intent data){
+        Uri uri = data.getData();
+        String imagePath = getImagePath(uri, null);
+        displayImage(imagePath);
+    }
+
+    private String getImagePath(Uri uri, String selection){
+        String path = null;
+        Cursor cursor = this.context.getContentResolver().query(uri, null, selection, null, null);
+        if(cursor != null){
+            if(cursor.moveToFirst()){
+                path = cursor.getString(cursor.getColumnIndex(MediaStore.Images.Media.DATA));
+            }
+            cursor.close();
+        }
+        return path;
+    }
+
+    private void displayImage(String imagePath){
+        if(imagePath != null){
+            File imgfile=new File(imagePath);
+            Log.d("picture", "!!!" + imgfile);
+            new Thread(()->{
+                boolean result= Http_setHeadImg.push(imgfile);
+                Log.e("RESULT", Boolean.toString(result));
+            }).start();
+        }else{
+            Toast.makeText(this.context, "faild to get image", Toast.LENGTH_SHORT).show();
+        }
+    }
+}
